added username restriction

+++ src/App/Controller/Slack.php
@@ -9,6 +9,7 @@ class Slack
     public function __construct(\App\Bluechilli\FireOrder $FireOrder)
     {
         $this->FireOrder = $FireOrder;
+        $this->commandlist = array('add', 'list','anyone','aim','fire');
 
         $storage = new \Flatbase\Storage\Filesystem('../src/db');
         $this->_db = new \Flatbase\Flatbase($storage);
@@ -42,12 +43,26 @@ class Slack
             // /shoot add username azimuth elevation (github)
             list($cmd, $username, $a, $e, $github) = explode (" ", $req['text']);
 
-            $this->_db->delete()->in('users')->where('slack', '==', $username)->execute();
-            $this->_db->insert()->in('users')
-                ->set(['slack' => $username, 'github' => $github, 'a' => $a, 'e' => $e])
-                ->execute();
+            if(in_array($username, $this->commandlist))
+            {
+                $payload = ["text" => $username . " is a protected command word and can't be used "];
+            }
+            else
+            {
 
-            $payload = ["text" => $username . " added at azimuth: ". $a. ", elevation: ". $e];
+                $this->_db->delete()->in('users')->where('slack', '==', $username)->execute();
+                $this->_db->insert()->in('users')
+                    ->set(['slack' => $username, 'github' => $github, 'a' => $a, 'e' => $e])
+                    ->execute();
+
+                $payload = ["text" => $username . " added at azimuth: ". $a. ", elevation: ". $e];
+            }
+        }
+
+        else if(substr($req['text'], 0, 3) == 'list')
+        {
+
+
         }
 
         else if(substr($req['text'], 0, 3) == 'aim')
